Edited Main code added material
I have included the datasheet for the 8-bit shift register. I have uploaded an image of the current setup that I have.  The 8 leds I have on the left are the outputs of the 8-bit shift register and the 5 pins on the right are the 5 control pins (only 4 are used). I have uploaded changes to the main code so that it will actually work. I have moved the boolean strint to bit string vi outside of the main program to its own subvi.