Binary Search continued
Added subVI that turns on and off a desired set of tubes.

Tried to work on a binary search using a for loop. I couldn't figure out how to continue with this method, so I am going to continue doing a long drawn-out bit of code using a lot of case structures.