Start of full brute force Dark Current Program.
Starting work on a full new Dark Current test baised on the brute force method. This will be in the "Full Dark Current VI (Brute Force)" folder. All new and "important" work will be in the "Brute Force Dark Current Primary Logic Body" VI. The center of this VI will contain all of the brains of the program inside, most likely, a for loop.