Added "Cleaned Up" version of brute force test
Here I created another directory to work separatly from the current project. I took the vis from the main brute force directory that we have been working on and copied it over. I decided to created a new vi called "Turn All Tubes ON-OFF" to simplify the code (it was repeated twice in the code). I then cleaned some stuff up. There are a few things that I will need to test if they work though.